Validate user can't create quote without text
Adds validation to ensure a quote can't be created withouth including
text. If a quote is attempted to be created without any text, the "New
Quote" page will be re-rendered with other fields pre-populated along
with a message for the user informing them of the error.

Feature specs have been created for the creation of a valid and invalid
quote.