Return to correct paginated page
After deleting a quote or after cancelling the editing of a quote,
displays a back button that takes the user back to the pagiated page
they were on.

Example, if the user is on the 5th paginated page and they delte a
quote, they are taken back to the 5th paginated page they were
originally on. If they cancel an edit, and choose to hit the back
button, they'll also be taken back to the 5th paginated page.