Add view helper to display a quote's "age"
This view helper will display the quotes age in number of days if it's
less than a week old. If it's over a week old, it'll simply display the
date the quote was created (i.e. "June 3, 2019"). If the quote is a
week old, we simply display "One week ago".